[Fix] 메인 및 마이페이지 수 230727 1750
메인페이지, 마이페이지에서 유저 상태관리 및 axios 통신 보강강

+++ back/B306/src/main/java/com/ssafy/B306/domain/template/TemplateService.java
@@ -1,12 +1,14 @@
 package com.ssafy.B306.domain.template;
 
 
-import com.ssafy.B306.domain.template.dto.TemplateRequestDto;
+import com.ssafy.B306.domain.security.JwtUtil;
 import com.ssafy.B306.domain.template.dto.TemplateSaveDto;
+import com.ssafy.B306.domain.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,14 +18,24 @@ import java.util.List;
 public class TemplateService {
 
     private final TemplateRepository templateRepository;
+    private final JwtUtil jwtUtil;
 
     // template 낱개 조회
-    public Template getTemplateAddress(TemplateRequestDto templateRequest){
-        return templateRepository.findByTemplateId(templateRequest.getTemplateId()).orElseThrow(() -> new RuntimeException("no template"));
+    public Template getTemplate(Long templateId){
+
+        Template template = templateRepository.findByTemplateId(templateId)
+                .orElseThrow(() -> new RuntimeException("no template"));
+
+        if(template.getTemplateDeleteDate() != null) {
+            throw new IllegalStateException("이미 삭제된 템플릿입니다.");
+        }
+
+        return template;
     }
 
     // 전체 template 조회
     public List<Template> getAllTemplate() {
+
         List<Template> temList = new ArrayList<>();
         temList = templateRepository.findAll();
         return temList;
@@ -31,27 +43,45 @@ public class TemplateService {
 
     // template 생성
     @Transactional
-    public Template addTemplate(TemplateSaveDto templateSaveDto) {
-        Template newTemp = templateSaveDto.toEntity(templateSaveDto);
-        //발생할만 한 문제가 뭐가있지?
+    public Template addTemplate(TemplateSaveDto templateSaveDto, HttpServletRequest request) {
 
-        return templateRepository.save(newTemp);
+        Long userPk = Long.parseLong(jwtUtil.parseClaims(request.getHeader("accessToken")).get("userPk").toString());
+        templateSaveDto.setUserPk(User.builder().userId(userPk).build());
+        Template template = templateSaveDto.toEntity(templateSaveDto);
+
+        return templateRepository.save(template);
     }
 
     // template 삭제
     @Transactional
-    public void deleteTemplate(Long templateId) {
-        Template template = templateRepository.findByTemplateId(templateId)
-                .orElseThrow(() -> new IllegalArgumentException("없는 템플릿 입니다."));
+    public void deleteTemplate(Long templateId, HttpServletRequest request) {
+
+        Long templateUserId = Long.parseLong(jwtUtil.parseClaims(request.getHeader("accessToken")).get("userPk").toString());
+
+        Template template = templateRepository.findByTemplateIdAndTemplateUserId(templateId, User.builder().userId(templateUserId).build())
+                .orElseThrow(() -> new IllegalArgumentException("해당 템플릿이 없습니다."));
+
+        if(template.getTemplateDeleteDate() != null) {
+            throw new IllegalStateException("이미 삭제된 템플릿입니다.");
+        }
 
         templateRepository.deleteById(template.getTemplateId());
     }
 
     // template 수정
     @Transactional
-    public void modifyTemplate(Long templateId, TemplateSaveDto templateSaveDto) {
-        Template template = templateRepository.findByTemplateId(templateId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 템플릿은 존재하지 않습니다."));
-        template.modifyTemplate(templateSaveDto);
+    public void modifyTemplate(Long templateId, TemplateSaveDto templateSaveDto, HttpServletRequest request) {
+
+        Long templateUserId = Long.parseLong(jwtUtil.parseClaims(request.getHeader("accessToken")).get("userPk").toString());
+
+        Template originalTemplate = templateRepository.findByTemplateIdAndTemplateUserId(templateId, User.builder().userId(templateUserId).build())
+                .orElseThrow(() -> new IllegalArgumentException("해당 템플릿이 없습니다."));
+
+        if(originalTemplate.getTemplateDeleteDate() != null) {
+            throw new IllegalStateException("이미 삭제된 템플릿입니다.");
+        }
+
+        originalTemplate.modifyTemplate(templateSaveDto);
+
     }
 }
